Subdued Brights, Added Shades
As the  title says, I changed the super bright tags to be more in line with actually not destroying eyeballs by clashing with the white text used in the tag system. I also added a spectrum of shades, blacks, greys & whites.

Although dunno if this requires more than just editing this one patch file, so lemme know!